Improve Makefile (#47)
Add a `uninstall` target and improve the `install` recipe.

The install recipe will now create the directories to hold the program
and man page. This is necessary if a custom prefix was used instead of
the default `/usr/local` where these directories most likely already
exist.

A custom prefix also implies root ownership of the executable may not be
required. Thus `DSTUSR` was set to the user running `make`. If the
default prefix is used, sudo is required and the user will be root
anyway -- nothing changed. However, with custom prefixes no attempt to
transfer ownership to root is made (which usually resulted in error).